+++ releng/release.sh
@@ -1,4 +1,4 @@
-#!/bin/bash 
+#!/bin/bash -eu
 
 ### Config
 DEV_BRANCH=develop
@@ -8,14 +8,14 @@ MAIN_BRANCH=master
 [ "$#" -ne "1" ] && echo "usage: $0 <new version>" && exit 1
 
 #warning
-echo "Warning: this script will delete new files and remove all changes in $DEV_BRANCH and $MAIN_BRANCH"
+echo "Warning: this script will delete ignored files and remove all changes in $DEV_BRANCH and $MAIN_BRANCH"
 read -p "Do you want to conitnue ? [NO/yes] " ANS
 LCANS=`echo "${ANS}" | tr '[:upper:]' '[:lower:]'`
 [ "${LCANS}" != "yes" ] && echo "Aborting." && exit 1
 
 NEW_VERSION=$1
 
-CURRENT_BRANCH=`git branch`
+CURRENT_BRANCH=$(cd `dirname $0` && echo `git branch`)
 ORIG_DIR=`pwd`
 DIR=$(cd `dirname $0` && echo `git rev-parse --show-toplevel`)
 TODAY_DATE=`date +%Y.%m.%d`
@@ -23,7 +23,7 @@ TODAY_DATE=`date +%Y.%m.%d`
 cd $DIR
 
 #move to dev branch and clean repo
-git stash
+git stash -u
 git checkout $DEV_BRANCH
 git reset --hard
 git clean -xdf
@@ -40,7 +40,7 @@ git checkout $MAIN_BRANCH
 git merge --no-ff $DEV_BRANCH -m "merge branch '$DEV_BRANCH' for new version $NEW_VERSION"
 git tag v$NEW_VERSION
 
-#move to snapshot version in develop
+#move to snapshot version in develop and push
 git checkout $DEV_BRANCH
 ./releng/update-version.sh $NEW_VERSION-SNAPSHOT
 cat release_notes.md | tail -n +3 > tmp
@@ -60,9 +60,19 @@ DFTools Changelog
 EOF
 cat tmp >> release_notes.md
 rm tmp
-
 git add -A
 git commit -m "[RELENG] Move to snapshot version"
+git push
 
+#deploy and push master (that is new version)
+git checkout master
+./releng/deploy.sh
+git push
+git push --tags
+
+#get back to original branch and restore work
+git checkout $CURRENT_BRANCH
+STASH_COUNT=`git stash list | wc -l`
+[ "$STASH_COUNT" != "0" ] && git stash pop
 #get back to original dir
 cd $ORIG_DIR